modified expanded nodes count and time limit cutoff

+++ src/GreedyBestFirstSearch.java
@@ -2,7 +2,7 @@ import java.util.List;
 
 public class GreedyBestFirstSearch extends Algorithm {
 	
-	public GreedyBestFirstSearch(double time, int startingNum, int targetNum, List<Action> actions) {
+	public GreedyBestFirstSearch(long time, int startingNum, int targetNum, List<Action> actions) {
 		super(time, startingNum, targetNum, actions);
 	}
